+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+* Fix internal accessibility error (#32603) when running `VIVIDUS: Run Stories` command
+
 ## [0.1.0] - 2023-02-03
 
 ### Added